+++ etna/models/nn/tft_native/__init__.py
@@ -1,4 +1,0 @@
-from etna import SETTINGS
-
-if SETTINGS.torch_required:
-    from etna.models.nn.tft_native.tft import TFTNativeModel